clean

+++ src/utils/types/query/methods.ts
@@ -27,6 +27,7 @@ import _ from 'lodash';
 import { Query } from './index';
 import { PObject } from '../object';
 import { Proto } from '../proto';
+import { PSchema } from '../schema';
 
 declare module './index' {
   export interface Query {
@@ -41,6 +42,10 @@ declare module './index' {
   }
 }
 
+const validateCLPs = (clps: PSchema.CLPs, acls: string[]) => {
+  return true;
+}
+
 export const queryMethods = (query: Query, proto: Proto, acls: string[]) => {
 
   const options = () => ({
@@ -49,15 +54,22 @@ export const queryMethods = (query: Query, proto: Proto, acls: string[]) => {
     ...query.options,
   });
 
+  const _validateCLPs = () => validateCLPs(
+    proto.schema[query.model]?.classLevelPermissions ?? {},
+    acls,
+  );
+
   const props = {
     count: {
       value: () => {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.count(options());
       },
     },
     then: {
       get() {
         const result = (async () => {
+          if (!_validateCLPs()) throw new Error('No permission');
           const array: PObject[] = [];
           for await (const obj of proto.storage.find(options())) array.push(obj);
           return array;
@@ -67,31 +79,37 @@ export const queryMethods = (query: Query, proto: Proto, acls: string[]) => {
     },
     [Symbol.asyncIterator]: {
       get() {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.find(options())[Symbol.asyncIterator];
       },
     },
     insert: {
       value: (attrs: any) => {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.insert(query.model, attrs);
       },
     },
     findOneAndUpdate: {
       value: (update: any) => {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.findOneAndUpdate(options(), update);
       },
     },
     findOneAndUpsert: {
       value: (update: any, setOnInsert: any) => {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.findOneAndUpsert(options(), update, setOnInsert);
       },
     },
     findOneAndDelete: {
       value: () => {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.findOneAndDelete(options());
       },
     },
     findAndDelete: {
       value: () => {
+        if (!_validateCLPs()) throw new Error('No permission');
         return proto.storage.findAndDelete(options());
       },
     },
